fix: fix issues of throw exception
