--- src/main/java/com/ugts/authentication/exception/authentication/AuthenticationException.java
@@ -1,14 +1,0 @@
-package com.ugts.authentication.exception.authentication;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthenticationException extends RuntimeException {
-    private AuthenticationErrorCode authenticationErrorCode;
-}
