--- src/main/java/com/ugts/authentication/exception/authentication/AuthenticationExceptionHandler.java
@@ -1,35 +1,0 @@
-package com.ugts.authentication.exception.authentication;
-
-import com.ugts.dto.ApiResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@Slf4j
-@ControllerAdvice
-public class AuthenticationExceptionHandler {
-
-    @ExceptionHandler(value = AuthenticationException.class)
-    ResponseEntity<ApiResponse> handleAuthenticationException(AuthenticationException e) {
-        AuthenticationErrorCode errorCode = e.getAuthenticationErrorCode();
-
-        return ResponseEntity.status(errorCode.getHttpStatusCode())
-                .body(ApiResponse.builder()
-                        .code(errorCode.getCode())
-                        .message(errorCode.getMessage())
-                        .build());
-    }
-
-    @ExceptionHandler(value = AccessDeniedException.class)
-    ResponseEntity<ApiResponse> handleAccessDeniedException(AccessDeniedException e) {
-        AuthenticationErrorCode errorCode = AuthenticationErrorCode.UNAUTHORIZED;
-
-        return ResponseEntity.status(errorCode.getHttpStatusCode())
-                .body(ApiResponse.builder()
-                        .code(errorCode.getCode())
-                        .message(errorCode.getMessage())
-                        .build());
-    }
-}
