--- src/main/java/com/ugts/authentication/exception/otp/OtpErrorCode.java
@@ -1,23 +1,0 @@
-package com.ugts.authentication.exception.otp;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.FieldDefaults;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public enum OtpErrorCode {
-    INVALID_OTP(1001, "Invalid OTP", HttpStatus.BAD_REQUEST),
-    OTP_EXPIRED(1002, "OTP has expired", HttpStatus.BAD_REQUEST),
-    OTP_NOT_EXISTED(1003, "OTP not existed", HttpStatus.BAD_REQUEST),
-    ;
-    int code;
-    String message;
-    HttpStatusCode httpStatusCode;
-}
