--- src/main/java/com/ugts/user/exception/UserException.java
@@ -1,14 +1,0 @@
-package com.ugts.user.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserException extends RuntimeException {
-    private UserErrorCode userErrorCode;
-}
