--- src/main/java/com/ugts/user/exception/UserExceptionHandler.java
@@ -1,22 +1,0 @@
-package com.ugts.user.exception;
-
-import com.ugts.dto.ApiResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@Slf4j
-@ControllerAdvice
-public class UserExceptionHandler {
-    @ExceptionHandler(value = UserException.class)
-    ResponseEntity<ApiResponse> handleAppException(UserException e) {
-        UserErrorCode errorCode = e.getUserErrorCode();
-
-        return ResponseEntity.status(errorCode.getHttpStatusCode())
-                .body(ApiResponse.builder()
-                        .code(errorCode.getCode())
-                        .message(errorCode.getMessage())
-                        .build());
-    }
-}
